+++ Airship/AirshipCore/Tests/ContactAPIClientTest.swift
@@ -152,5 +152,66 @@ class ContactAPIClientTest: XCTestCase {
         
         XCTAssertEqual(body as! NSDictionary, expectedBody as! NSDictionary)
     }
+    
+    func testUpdateInvalidAttribute() throws {
+        let date = Date()
+        let attributeUpdates = [
+            AttributeUpdate(attribute: "some-string", type: .set, value: nil, date: date),
+        ]
+        
+        let expectation = XCTestExpectation(description: "callback called")
+        contactAPIClient.update(identifier: "some-contact-id", tagGroupUpdates: [], attributeUpdates: attributeUpdates) { response, error in
+            XCTAssertNotNil(error)
+            expectation.fulfill()
+        }
+        
+        wait(for: [expectation], timeout: 10.0)
+    }
+    
+    func testUpdateMixValidInvalidAttributes() throws {
+        let date = Date()
+        let attributeUpdates = [
+            AttributeUpdate(attribute: "some-string", type: .set, value: nil, date: date),
+            AttributeUpdate.set(attribute: "some-string", value: "Hello", date: date),
+        ]
+        
+        let expectation = XCTestExpectation(description: "callback called")
+        contactAPIClient.update(identifier: "some-contact-id", tagGroupUpdates: [], attributeUpdates: attributeUpdates) { response, error in
+            XCTAssertEqual(response?.status, 200)
+            XCTAssertNil(error)
+            expectation.fulfill()
+        }
+        
+        wait(for: [expectation], timeout: 10.0)
+        
+        let request = self.session.lastRequest!
+        XCTAssertEqual("https://device-api.urbanairship.com/api/contacts/some-contact-id", request.url!.absoluteString)
+        
+        let body = try JSONSerialization.jsonObject(with: request.body!, options: [])
+        let formattedDate = Utils.isoDateFormatterUTCWithDelimiter().string(from: date)
+        let expectedBody : Any = [
+            "attributes" : [
+                [
+                    "action" : "set",
+                    "key" : "some-string",
+                    "timestamp" : formattedDate,
+                    "value" : "Hello"
+                ],
+            ]
+        ]
+        
+        XCTAssertEqual(body as! NSDictionary, expectedBody as! NSDictionary)
+    }
+    
+    
+    func testEmptyUpdates() throws {
+        let expectation = XCTestExpectation(description: "callback called")
+        contactAPIClient.update(identifier: "some-contact-id", tagGroupUpdates: [], attributeUpdates: []) { response, error in
+            XCTAssertNotNil(error)
+            expectation.fulfill()
+        }
+        
+        wait(for: [expectation], timeout: 10.0)
+    }
 
 }