+++ Airship/AirshipDebug/Source/Automation/TextInfoDetailViewController.swift
@@ -5,8 +5,8 @@ import UIKit
 #if canImport(AirshipCore)
 import AirshipCore
 import AirshipAutomation
-#elseif canImport(Airship)
-import Airship
+#elseif canImport(AirshipKit)
+import AirshipKit
 #endif
 
 /**