+++ Airship/AirshipDebug/Source/DeviceInfo/AddNamedUserAttributeTableViewController.swift
@@ -2,8 +2,8 @@
 
 #if canImport(AirshipCore)
 import AirshipCore
-#elseif canImport(Airship)
-import Airship
+#elseif canImport(AirshipKit)
+import AirshipKit
 #endif
 
 class AddNamedUserAttributeTableViewController: AddAttributeTableViewController {